core/pyproject.toml

[project]
name = "dvt-ce"
dynamic = ["version"]
description = "Data Virtualization Tool - Federated queries across data sources"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "Apache-2.0"
license-files = { globs = ["LICENSE", "NOTICE"] }
keywords = []
authors = [
    { name = "Hesham Badawi" },
]
maintainers = [
    { name = "Meta Thinking Dev Team" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    # ----
    # dbt-core uses these packages deeply, throughout the codebase, and there have been breaking changes in past patch releases (even though these are major-version-one).
    # Pin to the patch or minor version, and bump in each new minor version of dbt-core.
    "agate>=1.7.0,<1.10",
    "Jinja2>=3.1.3,<4",
    "mashumaro[msgpack]>=3.9,<3.15",
    # ----
    # dbt-core uses these packages in standard ways. Pin to the major version, and check compatibility
    # with major versions in each new minor version of dbt-core.
    "click>=8.2.0,<9.0",
    "jsonschema>=4.19.1,<5.0",
    "networkx>=2.3,<4.0",
    "protobuf>=6.0,<7.0",
    "requests<3.0.0",  # should match dbt-common
    "snowplow-tracker>=1.0.2,<1.1",
    # ----
    # These packages are major-version-0. Keep upper bounds on upcoming minor versions (which could have breaking changes)
    # and check compatibility / bump in each new minor version of dbt-core.
    "pathspec>=0.9,<0.13",
    "sqlparse>=0.5.0,<0.5.5",
    # ----
    # These are major-version-0 packages also maintained by dbt-labs.
    # Accept patches but avoid automatically updating past a set minor version range.
    "dbt-extractor>=0.5.0,<=0.6",
    "dbt-semantic-interfaces>=0.9.0,<0.10",
    # Minor versions for these are expected to be backwards-compatible
    "dbt-common>=1.37.2,<2.0",
    "dvt-adapters>=0.1.0,<2.0",
    "dbt-protos>=1.0.405,<2.0",
    # DVT-specific: pyspark is not a dependency; dvt sync installs it per project.
    "sqlglot>=20.0.0",
    "duckdb>=0.9.0",
    "pyarrow>=14.0.0",
    "pydantic<3",
    # ----
    # Expect compatibility with all new versions of these packages, so lower bounds only.
    "packaging>20.9",
    "pytz>=2015.7",
    "pyyaml>=6.0",
    "daff>=1.3.46",
    "typing-extensions>=4.4",
]

[project.urls]
Homepage = "https://github.com/heshamh96/dvt-ce"
Repository = "https://github.com/heshamh96/dvt-ce.git"
Issues = "https://github.com/heshamh96/dvt-ce/issues"
Changelog = "https://github.com/heshamh96/dvt-ce/blob/dev/CHANGELOG.md"

[project.scripts]
dvt = "dvt.cli.main:cli"

[tool.hatch.version]
path = "dvt/__version__.py"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "hypothesis>=6.151.5",
    "twine>=6.2.0",
]
